refactor(brand): isolate server-only og exports
